fix header footer
